requirements.txt
bs4
langchain_community
langchain_openai
python-dotenv
flask
gunicorn